requirements.txt
langchain
langchain-openai
langfuse>=3,<4
python-dotenv
ulid-py